refactor: Hide information in small screens
This commit refactors the layout in the frame.astro file to hide the "Information", "RecraftCPH", "Nyhedsbrev", and "Forhandler" sections on small screens. It adds the "hide-on-small" class to the corresponding paragraphs to apply the CSS rule that sets their display property to none. This change improves the responsiveness and usability of the website.